RaftedValue.force_remove_member/2 no longer checks whether a leader exists or not
This is a slightly breaking change.
Alghouth it looks a bit safer for `force_remove_member/2` to reject if a leader exists,
it turns out that this actually prevents us from completely recovering from majority failures,
as once a leader is elected we cannot remove a member even using `force_remove_member/2`.
Now `force_remove_member/2` makes a living member to forget about a (probably dead) member
regardless of existence of a leader.